6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ long int d;
 long long int f;
 float e;
 
-printf("Size of char: %lu byte(s)\n", (unsigned long) sizeof(a));
-printf("Size of int: %lu byte(s)\n", (unsigned long) sizeof(b));
-printf("Size of long int: %lu byte(s)\n", (unsigned long) sizeof(d));
-printf("Size of long long int: %lu byte(s)\n", (unsigned long) sizeof(f));
-printf("Size of float: %lu byte(s)\n", (unsigned long) sizeof(e));
+printf("Size of char: %lu byte(s)\n", sizeof(a));
+printf("Size of int: %lu byte(s)\n", sizeof(b));
+printf("Size of long int: %lu byte(s)\n",sizeof(d));
+printf("Size of long long int: %lu byte(s)\n", sizeof(f));
+printf("Size of float: %lu byte(s)\n", sizeof(e));
 return (0);
 }
